\name{print_equilibria}
\alias{print_equilibria}
\title{Print Nash Equilibria}
\description{
Pretty-print a list of Nash equilibria.
}
\usage{
print_equilibria(equilibria, game)
}
\arguments{
\item{equilibria}{A list of equilibria from nash_equilibria.}
\item{game}{A NormalFormGame object.}
}
\value{
Invisibly returns equilibria.
}
