\name{support_enumeration}
\alias{support_enumeration}
\title{Support Enumeration for Nash Equilibria}
\description{
Find all Nash equilibria of a 2-player game using support enumeration.
}
\usage{
support_enumeration(game, tol = 1e-10)
}
\arguments{
\item{game}{A NormalFormGame object (2-player).}
\item{tol}{Numeric tolerance.}
}
\value{
A list of equilibria.
}
